+++ CHANGELOG.md
@@ -5,6 +5,8 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## Unreleased
+### Added
+- Adds a "VIDEO" tag to DAM video thumbnails
 ### Fixed
 - Cleaned up some exception handing in the DAM fieldtype